// angular_src/src/app/app.api.ts
export class CropVm implements ICropVm {
    createdOn?: Date | null;
    updatedOn?: Date | null;
    _id?: string | null;
    name: string;
    variety: string[];
    pricePerPound: number;

    constructor(data?: ICropVm) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
        if (!data) {
            this.variety = [];
        }
    }

    init(data?: any) {
        if (data) {
            this.createdOn = data["createdOn"] ? new Date(data["createdOn"].toString()) : <any>null;
            this.updatedOn = data["updatedOn"] ? new Date(data["updatedOn"].toString()) : <any>null;
            this._id = data["_id"] !== undefined ? data["_id"] : <any>null;
            this.name = data["name"] !== undefined ? data["name"] : <any>null;
            if (data["variety"] && data["variety"].constructor === Array) {
                this.variety = [];
                for (let item of data["variety"])
                    this.variety.push(item);
            }
            this.pricePerPound = data["pricePerPound"] !== undefined ? data["pricePerPound"] : <any>null;
        }
    }

    static fromJS(data: any): CropVm {
        data = typeof data === 'object' ? data : {};
        let result = new CropVm();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["createdOn"] = this.createdOn ? this.createdOn.toISOString() : <any>null;
        data["updatedOn"] = this.updatedOn ? this.updatedOn.toISOString() : <any>null;
        data["_id"] = this._id !== undefined ? this._id : <any>null;
        data["name"] = this.name !== undefined ? this.name : <any>null;
        if (this.variety && this.variety.constructor === Array) {
            data["variety"] = [];
            for (let item of this.variety)
                data["variety"].push(item);
        }
        data["pricePerPound"] = this.pricePerPound !== undefined ? this.pricePerPound : <any>null;
        return data;
    }
}

export interface ICropVm {
    createdOn?: Date | null;
    updatedOn?: Date | null;
    _id?: string | null;
    name: string;
    variety: string[];
    pricePerPound: number;
}

export class HarvesterVm implements IHarvesterVm {
    createdOn?: Date | null;
    updatedOn?: Date | null;
    _id?: string | null;
    firstName: string;
    lastName: string;

    constructor(data?: IHarvesterVm) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.createdOn = data["createdOn"] ? new Date(data["createdOn"].toString()) : <any>null;
            this.updatedOn = data["updatedOn"] ? new Date(data["updatedOn"].toString()) : <any>null;
            this._id = data["_id"] !== undefined ? data["_id"] : <any>null;
            this.firstName = data["firstName"] !== undefined ? data["firstName"] : <any>null;
            this.lastName = data["lastName"] !== undefined ? data["lastName"] : <any>null;
        }
    }

    static fromJS(data: any): HarvesterVm {
        data = typeof data === 'object' ? data : {};
        let result = new HarvesterVm();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["createdOn"] = this.createdOn ? this.createdOn.toISOString() : <any>null;
        data["updatedOn"] = this.updatedOn ? this.updatedOn.toISOString() : <any>null;
        data["_id"] = this._id !== undefined ? this._id : <any>null;
        data["firstName"] = this.firstName !== undefined ? this.firstName : <any>null;
        data["lastName"] = this.lastName !== undefined ? this.lastName : <any>null;
        return data;
    }
}
// ...
export class EntryVm implements IEntryVm {
    createdOn?: Date | null;
    updatedOn?: Date | null;
    _id?: string | null;
    crop: CropVm;
    pounds: number;
    priceTotal: number;
    harvester: HarvesterVm;
    comments: string;
    recipient: OrganizationVm;
    selectedVariety: string;

    constructor(data?: IEntryVm) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
        if (!data) {
            this.crop = new CropVm();
            this.harvester = new HarvesterVm();
            this.recipient = new OrganizationVm();
        }
    }

    init(data?: any) {
        if (data) {
            this.createdOn = data["createdOn"] ? new Date(data["createdOn"].toString()) : <any>null;
            this.updatedOn = data["updatedOn"] ? new Date(data["updatedOn"].toString()) : <any>null;
            this._id = data["_id"] !== undefined ? data["_id"] : <any>null;
            this.crop = data["crop"] ? CropVm.fromJS(data["crop"]) : new CropVm();
            this.pounds = data["pounds"] !== undefined ? data["pounds"] : <any>null;
            this.priceTotal = data["priceTotal"] !== undefined ? data["priceTotal"] : <any>null;
            this.harvester = data["harvester"] ? HarvesterVm.fromJS(data["harvester"]) : new HarvesterVm();
            this.comments = data["comments"] !== undefined ? data["comments"] : <any>null;
            this.recipient = data["recipient"] ? OrganizationVm.fromJS(data["recipient"]) : new OrganizationVm();
            this.selectedVariety = data["selectedVariety"] !== undefined ? data["selectedVariety"] : <any>null;
        }
    }

    static fromJS(data: any): EntryVm {
        data = typeof data === 'object' ? data : {};
        let result = new EntryVm();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["createdOn"] = this.createdOn ? this.createdOn.toISOString() : <any>null;
        data["updatedOn"] = this.updatedOn ? this.updatedOn.toISOString() : <any>null;
        data["_id"] = this._id !== undefined ? this._id : <any>null;
        data["crop"] = this.crop ? this.crop.toJSON() : <any>null;
        data["pounds"] = this.pounds !== undefined ? this.pounds : <any>null;
        data["priceTotal"] = this.priceTotal !== undefined ? this.priceTotal : <any>null;
        data["harvester"] = this.harvester ? this.harvester.toJSON() : <any>null;
        data["comments"] = this.comments !== undefined ? this.comments : <any>null;
        data["recipient"] = this.recipient ? this.recipient.toJSON() : <any>null;
        data["selectedVariety"] = this.selectedVariety !== undefined ? this.selectedVariety : <any>null;
        return data;
    }
}

export interface IEntryVm {
    createdOn?: Date | null;
    updatedOn?: Date | null;
    _id?: string | null;
    crop: CropVm;
    pounds: number;
    priceTotal: number;
    harvester: HarvesterVm;
    comments: string;
    recipient: OrganizationVm;
    selectedVariety: string;
}
// ...
export class FarmVm implements IFarmVm {
    createdOn?: Date | null;
    updatedOn?: Date | null;
    _id?: string | null;
    name: string;
    lat: number;
    lng: number;

    constructor(data?: IFarmVm) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.createdOn = data["createdOn"] ? new Date(data["createdOn"].toString()) : <any>null;
            this.updatedOn = data["updatedOn"] ? new Date(data["updatedOn"].toString()) : <any>null;
            this._id = data["_id"] !== undefined ? data["_id"] : <any>null;
            this.name = data["name"] !== undefined ? data["name"] : <any>null;
            this.lat = data["lat"] !== undefined ? data["lat"] : <any>null;
            this.lng = data["lng"] !== undefined ? data["lng"] : <any>null;
        }
    }

    static fromJS(data: any): FarmVm {
        data = typeof data === 'object' ? data : {};
        let result = new FarmVm();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["createdOn"] = this.createdOn ? this.createdOn.toISOString() : <any>null;
        data["updatedOn"] = this.updatedOn ? this.updatedOn.toISOString() : <any>null;
        data["_id"] = this._id !== undefined ? this._id : <any>null;
        data["name"] = this.name !== undefined ? this.name : <any>null;
        data["lat"] = this.lat !== undefined ? this.lat : <any>null;
        data["lng"] = this.lng !== undefined ? this.lng : <any>null;
        return data;
    }
}

export interface IFarmVm {
    createdOn?: Date | null;
    updatedOn?: Date | null;
    _id?: string | null;
    name: string;
    lat: number;
    lng: number;
}

export class HarvestVm implements IHarvestVm {
    createdOn?: Date | null;
    updatedOn?: Date | null;
    _id?: string | null;
    farm: FarmVm;
    entries: EntryVm[];

    constructor(data?: IHarvestVm) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
        if (!data) {
            this.farm = new FarmVm();
            this.entries = [];
        }
    }

    init(data?: any) {
        if (data) {
            this.createdOn = data["createdOn"] ? new Date(data["createdOn"].toString()) : <any>null;
            this.updatedOn = data["updatedOn"] ? new Date(data["updatedOn"].toString()) : <any>null;
            this._id = data["_id"] !== undefined ? data["_id"] : <any>null;
            this.farm = data["farm"] ? FarmVm.fromJS(data["farm"]) : new FarmVm();
            if (data["entries"] && data["entries"].constructor === Array) {
                this.entries = [];
                for (let item of data["entries"])
                    this.entries.push(EntryVm.fromJS(item));
            }
        }
    }

    static fromJS(data: any): HarvestVm {
        data = typeof data === 'object' ? data : {};
        let result = new HarvestVm();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["createdOn"] = this.createdOn ? this.createdOn.toISOString() : <any>null;
        data["updatedOn"] = this.updatedOn ? this.updatedOn.toISOString() : <any>null;
        data["_id"] = this._id !== undefined ? this._id : <any>null;
        data["farm"] = this.farm ? this.farm.toJSON() : <any>null;
        if (this.entries && this.entries.constructor === Array) {
            data["entries"] = [];
            for (let item of this.entries)
                data["entries"].push(item.toJSON());
        }
        return data;
    }
}

export interface IHarvestVm {
    createdOn?: Date | null;
    updatedOn?: Date | null;
    _id?: string | null;
    farm: FarmVm;
    entries: EntryVm[];
}